+++ src/Input/RequestStringInput.php
@@ -26,23 +26,20 @@ use Art4\JsonApiClient\Utils\Helper;
  */
 final class RequestStringInput implements RequestInput
 {
+    use StringInputTrait;
+
     private $rawString;
 
     /**
      * Set the input
      *
      * @param string $string
+     *
+     * @throws InputException if $string is not a string
      */
     public function __construct($string)
     {
-        if (! is_string($string)) {
-            throw new \Exception(sprintf(
-                '$string must be a string, "%s" given.',
-                gettype($string)
-            ));
-        }
-
-        $this->string = $string;
+        $this->rawString = $this->prepareString($string);
     }
 
     /**
@@ -51,10 +48,12 @@ final class RequestStringInput implements RequestInput
      * This should be a native PH stdClass object, so Manager could
      * iterate over all public attributes
      *
+     * @throws InputException if somethin went wrong with the input
+     *
      * @return stdClass
      */
     public function getAsObject()
     {
-        return Helper::decodeJson($this->string);
+        return $this->decodeJson($this->rawString);
     }
 }